moved bbc into scraper file
